+++ apps/users/templates/users/login.html
@@ -70,7 +70,9 @@
                         "
                       >
                         <a class="small" href="#">Forgot Password?</a>
-                        <a class="btn btn-primary" href="/jomi-passes">Login</a>
+                        <a class="btn btn-primary" href="/passwords-list"
+                          >Login</a
+                        >
                       </div>
                     </form>
                   </div>
