+++ apps/users/templates/users/register.html
@@ -92,7 +92,7 @@
                       </div>
                       <div class="mt-4 mb-0">
                         <div class="d-grid">
-                          <a class="btn btn-primary btn-block" href="/"
+                          <a class="btn btn-primary btn-block" href="/login"
                             >Create Account</a
                           >
                         </div>
@@ -101,7 +101,7 @@
                   </div>
                   <div class="card-footer text-center py-3">
                     <div class="small">
-                      <a href="/">Have an account? Go to login</a>
+                      <a href="/login">Have an account? Go to login</a>
                     </div>
                   </div>
                 </div>
