+++ templates/layout.html
@@ -33,7 +33,7 @@
         {% include '/partials/footer.html' %}
       </div>
     </div>
-    <script src="{{ url_for('static', filename='js/scripts..js')}}"></script>
+    <script src="{{ url_for('static', filename='js/scripts.js')}}"></script>
     <script src="{{ url_for('static', filename='js/bootstrap.min.js')}}"></script>
     <script src="{{ url_for('static', filename='js/Chart.js')}}"></script>
     <script src="{{ url_for('static', filename='js/simple-datatables.js')}}"></script>
